Выносит логику отправки и загрузки статистики в app.js
